+++ src/BabelFish.Reports/DataActors/Excel/ResultListExcel.cs
@@ -1,233 +1,138 @@
 using OfficeOpenXml;
 using Scopos.BabelFish.APIClients;
-using Scopos.BabelFish.DataActors.OrionMatch;
-using Scopos.BabelFish.DataModel.Definitions;
+using Scopos.BabelFish.DataActors.ResultListFormatter;
 using Scopos.BabelFish.DataModel.OrionMatch;
-using Scopos.BabelFish.Helpers;
 
 namespace Scopos.BabelFish.DataActors.Excel {
-    public class ResultListExcel : ExcelGenerator {
+    public class ResultListExcel : ExcelGenerator<ResultList> {
 
-        private ResultList ResultList { get; set; } = null;
-        private ResultList PassedResultList { get; set; } = null;
+        public IRLIFList? TheList { get; private set; }
 
-        private List<string> resultEventsToInclude = new List<string>();
+        public List<ResultListIntermediateFormatted>? WorksheetData { get; private set; }
 
-        private List<string> resultEventsNamesToInclude = new List<string>();
-
-        private EventComposite eventTree = null;
-
-        private ExcelWorksheet? Worksheet1 = null;
-        private ExcelWorksheet? Worksheet2 = null;
-
-        public ResultListExcel( ResultList resultList ) {
-            ResultList = resultList;
+        /// <summary>
+        /// Private constructor, so users have to use the InitializeAsync method.
+        /// </summary>
+        private ResultListExcel() {
         }
 
-        public async Task InitalizeAsync() {
-            await CreateEventTreeAsync( ResultList );
-        }
+        /// <summary>
+        /// Factory method to construct a new instance of ResultListExcel based on the passed in ResultList, and
+        /// optionall SquaddingList.
+        /// </summary>
+        /// <param name="resultList"></param>
+        /// <param name="squaddingList"></param>
+        /// <exception cref="ArgumentNullException" >Thrown when the passed in resultList parameter is null.</exception>
+        /// <returns></returns>
+        public static async Task<ResultListExcel> FactoryAsync( ResultList resultList, SquaddingList? squaddingList = null ) {
+            if (resultList is null)
+                throw new ArgumentNullException( "The argument resultList may not be null. " );
 
-        private async Task CreateEventTreeAsync( ResultList resultList ) {
-            #region CreateEventTree
+            var rle = new ResultListExcel();
+            rle.TheList = resultList;
 
-            SetName cofSetName;
-            CourseOfFire? courseOfFireDefinition = null;
-            string courseOfFireCommonName;
-            if (resultList.CourseOfFireDef != "v1.0:orion:unknown" && SetName.TryParse( resultList.CourseOfFireDef, out cofSetName )) {
-                courseOfFireDefinition = await DefinitionCache.GetCourseOfFireDefinitionAsync( cofSetName );
-                courseOfFireCommonName = courseOfFireDefinition.CommonName;
+            rle.WorksheetData = new List<ResultListIntermediateFormatted>();
 
-                eventTree = EventComposite.GrowEventTree( courseOfFireDefinition );
-            }
+            //For worksheet 1, use the defined RESULT LIST FORMAT for the passed in Result List
+            var resultListFormatSetName = await ResultListFormatFactory.FACTORY.GetResultListFormatSetNameAsync( resultList );
+            var rlfWs1 = await DefinitionCache.GetResultListFormatDefinitionAsync( resultListFormatSetName );
 
+            ResultListIntermediateFormatted worksheet1 = new ResultListIntermediateFormatted( resultList, rlfWs1, null );
+            await worksheet1.InitializeAsync();
+            rle.WorksheetData.Add( worksheet1 );
 
-            if (resultList.Projected) {
-                CompareResultByRank.CompareMethod method = CompareResultByRank.CompareMethod.RANK_ORDER;
-                SortBy sortBy = SortBy.ASCENDING;
-                CompareResultByRank compareResultByRank = new CompareResultByRank( method, sortBy );
-                resultList.Items.Sort( compareResultByRank );
-
-            }
-
-
-            //Add to the lists to display, if there are any results
-            if (resultList.Items.Count > 0 && eventTree != null) {
-                //add only the EVENT, STAGE and SERIES events in the event tree to be shown on the sheet
-                foreach (var item in eventTree.GetEvents( true, true, true, true, true, false )) {
-                    resultEventsToInclude.Add( item.EventName );
-                    resultEventsNamesToInclude.Add( item.EventName + " - DEC" );
-                    resultEventsNamesToInclude.Add( item.EventName + " - INT" );
-                    resultEventsNamesToInclude.Add( item.EventName + " - X" );
-                }
-            } else {
-                //we need to throw some error here, though I don't think this will ever be hit if the RL doesn't load.
-                ;
-            }
-            #endregion
-        }
-
-        public override string GenerateExcel( string? filePath = null ) {
-            using (var package = new ExcelPackage()) {
-                Worksheet1 = package.Workbook.Worksheets.Add( this.ResultList.Status + " ResultList" );
-                List<string> header = new List<string>();
-                if (this.ResultList.Team) {
-                    //Creates 2 worksheets, #one for teams only, #two for individuals only.
-                    Worksheet2 = package.Workbook.Worksheets.Add( "Individuals" );
-                    List<string>? header2 = new List<string>();
-
-                    header.Add( "Name" );
-                    header.Add( "Rank" );
-                    List<string> staticVars = new List<string>();
-                    staticVars.AddRange( header );
-                    header.AddRange( resultEventsNamesToInclude );
-                    Worksheet1 = PopulateHeader( Worksheet1, header );
-                    List<ExcelResultEvent> teams = new List<ExcelResultEvent>();
-                    foreach (var team in this.ResultList.Items) {
-                        Dictionary<string, string> data = new Dictionary<string, string>();
-                        data.Add( header[0], team.Participant.DisplayName );
-                        data.Add( header[1], team.Rank.ToString() );
-
-                        ExcelResultEvent teamExcel = new ExcelResultEvent( team, data );
-                        teams.Add( teamExcel );
-                    }
-                    Worksheet1 = PopulateData( Worksheet1, teams, resultEventsToInclude, staticVars );
-
-                    header2.Add( "Name" );
-                    header2.Add( "Competitor Number" );
-                    header2.Add( "Rank" );
-                    header2.Add( "Team" );
-                    List<string> staticVars2 = new List<string>();
-                    staticVars2.AddRange( header2 );
-                    header2.AddRange( resultEventsNamesToInclude );
-                    Worksheet2 = PopulateHeader( Worksheet2, header2 );
-                    //make a long list of all team members and pass it to this one.
-                    // how get what team they are in??
-                    //  could populate one team at a time? that could work....
-                    //   would need to keep track of start/stop and team name
-                    List<ExcelResultEvent> individuals = new List<ExcelResultEvent>();
-                    foreach (var team in ResultList.Items) {
-                        if (team.TeamMembers != null) {
-                            foreach (var teamMember in team.TeamMembers) {
-                                Dictionary<string, string> data = new Dictionary<string, string>();
-                                data.Add( header2[0], teamMember.Participant.DisplayName );
-                                data.Add( header2[1], teamMember.Participant.CompetitorNumber );
-                                data.Add( header2[2], team.Rank.ToString() ); // choosing to use team rank here, though they are in order of score by team
-                                data.Add( header2[3], team.Participant.DisplayName );
-
-                                ExcelResultEvent teamExcel = new ExcelResultEvent( teamMember, data );
-                                individuals.Add( teamExcel );
-                            }
-                        }
-                    }
-
-                    Worksheet2 = PopulateData( Worksheet2, individuals, resultEventsToInclude, staticVars2 );
-
-                    Worksheet1.Cells[Worksheet1.Dimension.Address].AutoFitColumns();
-                    Worksheet1.View.FreezePanes( 2, 1 );
-
-                    Worksheet2.Cells[Worksheet2.Dimension.Address].AutoFitColumns();
-                    Worksheet2.View.FreezePanes( 2, 1 );
+            //For workseet 2, use a dynamically created Essential Data File RESULT LIST FORMAT
+            var essentialDataFile = new DynamicEssentialDataFile();
+            var rlfWs2 = await essentialDataFile.GenerateAsync( resultList );
+            ResultListIntermediateFormatted worksheet2 = new ResultListIntermediateFormatted( resultList, rlfWs2, null );
+            await worksheet2.InitializeAsync();
+            //Add in squadding if the result list is future, intermediate, or unofficial
+            if (resultList.Status != ResultStatus.OFFICIAL) {
+                //If the user passed in a squadding list, we can use it. Otherwise, try and load it from REST API
+                if (squaddingList is not null) {
+                    worksheet2.LoadSquaddingList( squaddingList );
+                    worksheet2.RefreshAllRowsParticipantAttributeFields();
                 } else {
-                    //creates only one worksheet for individuals
-                    header.Add( "Name" );
-                    header.Add( "Competitor Number" );
-                    header.Add( "Rank" );
-                    List<string> staticVars = new List<string>();
-                    staticVars.AddRange( header );
-                    header.AddRange( resultEventsNamesToInclude );
-                    Worksheet1 = PopulateHeader( Worksheet1, header );
-                    List<ExcelResultEvent> individuals = new List<ExcelResultEvent>();
-                    foreach (var person in ResultList.Items) {
-                        Dictionary<string, string> data = new Dictionary<string, string>();
-                        data.Add( header[0], person.Participant.DisplayName );
-                        data.Add( header[1], person.Participant.CompetitorNumber );
-                        data.Add( header[2], person.Rank.ToString() );
+                    await worksheet2.LoadSquaddingListAsync();
+                }
+            }
+            rle.WorksheetData.Add( worksheet2 );
 
-                        ExcelResultEvent PersonExcel = new ExcelResultEvent( person, data );
-                        individuals.Add( PersonExcel );
-                    }
-                    Worksheet1 = PopulateData( Worksheet1, individuals, resultEventsToInclude, staticVars );
+            return rle;
+        }
 
-                    Worksheet1.Column( 3 ).Style.Numberformat.Format = "0.00";
-                    Worksheet1.Cells[Worksheet1.Dimension.Address].AutoFitColumns();
-                    Worksheet1.View.FreezePanes( 2, 1 );
+        public static async Task<ResultListExcel> FactoryAsync( SquaddingList squaddingList ) {
+            if (squaddingList is null)
+                throw new ArgumentNullException( "The argument squaddingList may not be null." );
+
+            var rle = new ResultListExcel();
+            rle.TheList = squaddingList;
+
+            rle.WorksheetData = new List<ResultListIntermediateFormatted>();
+
+            var dynamicSquadding = new DynamicSquadding();
+            var squaddingResultListFormatDefinition = await dynamicSquadding.GenerateAsync( squaddingList );
+            var worksheet1 = new ResultListIntermediateFormatted( squaddingList, squaddingResultListFormatDefinition, null );
+            await worksheet1.InitializeAsync();
+            rle.WorksheetData.Add( worksheet1 );
+
+            return rle;
+        }
+
+        /// <inheritdoc />
+        public override byte[] GenerateExcel( string? filePath = null ) {
+            using (var package = new ExcelPackage()) {
+                foreach (var rlif in this.WorksheetData) {
+                    var worksheet = package.Workbook.Worksheets.Add( $"{this.TheList.Name}: {rlif.ResultListFormat.CommonName}" );
+                    PopulateWorksheet( worksheet, rlif );
                 }
 
                 if (filePath != null) {
                     package.SaveAs( filePath );
                 }
+
                 // Setting the buffer to hold the bytes of the Excel file
                 var buffer = package.GetAsByteArray();
 
+                return buffer;
+
+                /*
                 //I think I should simply return the buffer, then the website handles displaying it.
                 return Convert.ToBase64String( buffer );
+                */
             }
         }
 
-        private ExcelWorksheet PopulateHeader( ExcelWorksheet worksheet, List<string> headers ) {
-            int currentCol = 1;
-            foreach (string colName in headers) {
-                worksheet.Cells[1, currentCol].Value = colName;
-                worksheet.Cells[1, currentCol].Style.Font.Bold = true;
-                currentCol++;
+        protected static void PopulateWorksheet( ExcelWorksheet worksheet, ResultListIntermediateFormatted rlif ) {
+            int columnIndex = 1;
+            int rowIndex = 1;
+
+            foreach (var headerCell in rlif.GetShownHeaderRow()) {
+                worksheet.Cells[rowIndex, columnIndex].Value = headerCell.Text;
+                worksheet.Cells[rowIndex, columnIndex].Style.Font.Bold = true;
+                columnIndex++;
             }
-            return worksheet;
-        }
 
-
-        private ExcelWorksheet PopulateData( ExcelWorksheet worksheet, List<ExcelResultEvent> excelResultEvents, List<string> headers, List<string> staticVars ) {
-            //static column names are different than the ones expected in the ResultEvent
-            // Adding Data
-            int currentCol = 1;
-            var currentRow = 2; // Starting from the second row because the first row is for headers
-            foreach (var resultEvent in excelResultEvents) {
-                //fill in the name/rank/compnum/that stuff here
-                foreach (var data in staticVars) {
-                    worksheet.Cells[currentRow, currentCol].Value = resultEvent.Data[data];
-                    currentCol++;
-                }
-                foreach (var resultEventName in headers) {
-                    if (resultEvent.ResultEvent.EventScores.ContainsKey( resultEventName )) {
-                        // decimal
-                        worksheet.Cells[currentRow, currentCol].Value = resultEvent.ResultEvent.EventScores[resultEventName].Score.D;
-                        currentCol++;
-
-                        // integer
-                        worksheet.Cells[currentRow, currentCol].Value = resultEvent.ResultEvent.EventScores[resultEventName].Score.I;
-                        currentCol++;
-
-                        // inners
-                        worksheet.Cells[currentRow, currentCol].Value = resultEvent.ResultEvent.EventScores[resultEventName].Score.X;
-                        currentCol++;
-                    } else {
-                        // decimal
-                        worksheet.Cells[currentRow, currentCol].Value = 0.0;
-                        currentCol++;
-
-                        // integer
-                        worksheet.Cells[currentRow, currentCol].Value = 0;
-                        currentCol++;
-
-                        // inners
-                        worksheet.Cells[currentRow, currentCol].Value = 0;
-                        currentCol++;
+            foreach (var multilineRow in rlif.Rows) {
+                foreach (var row in multilineRow) {
+                    rowIndex++;
+                    columnIndex = 1;
+                    foreach (var cell in row.GetShownRow()) {
+                        worksheet.Cells[rowIndex, columnIndex].Value = ParseSmart( cell.Text );
+                        columnIndex++;
                     }
                 }
-                currentRow++;
-                currentCol = 1;
             }
-            return worksheet;
         }
-    }
-    public class ExcelResultEvent {
-        public ResultEvent ResultEvent { get; set; }
 
-        public Dictionary<string, string> Data = new Dictionary<string, string>();
+        protected static object ParseSmart( string input ) {
+            if (int.TryParse( input, out int intValue ))
+                return intValue;
 
-        public ExcelResultEvent( ResultEvent resultEvent, Dictionary<string, string> data ) {
-            ResultEvent = resultEvent;
-            Data = data;
+            if (float.TryParse( input, out float floatValue ))
+                return floatValue;
+
+            return input;
         }
     }
 }